requirements.txt
Django
pytz
python-dotenv